Moving response exception to lib directory
